+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fixed error lint with eslint ([]())
+
 ### Changed
 
 -   Migrated tslint to eslint. ([#331](https://github.com/reyesoft/ngx-jsonapi/pull/331))